+++ src/main/java/com/atlassian/db/replica/api/DualConnection.java
@@ -1,6 +1,6 @@
 package com.atlassian.db.replica.api;
 
-import com.atlassian.db.replica.api.circuitbreaker.BreakerState;
+import com.atlassian.db.replica.internal.circuitbreaker.BreakerState;
 import com.atlassian.db.replica.api.reason.Reason;
 import com.atlassian.db.replica.internal.state.NoOpStateListener;
 import com.atlassian.db.replica.internal.ForwardCall;
@@ -16,7 +16,7 @@ import com.atlassian.db.replica.internal.circuitbreaker.BreakerHandler;
 import com.atlassian.db.replica.spi.ConnectionProvider;
 import com.atlassian.db.replica.spi.DatabaseCall;
 import com.atlassian.db.replica.spi.ReplicaConsistency;
-import com.atlassian.db.replica.spi.circuitbreaker.CircuitBreaker;
+import com.atlassian.db.replica.internal.circuitbreaker.CircuitBreaker;
 import com.atlassian.db.replica.internal.state.StateListener;
 
 import java.sql.Array;
@@ -540,7 +540,7 @@ public final class DualConnection implements Connection {
             return this;
         }
 
-        public DualConnection.Builder circuitBreaker(CircuitBreaker circuitBreaker) {
+        DualConnection.Builder circuitBreaker(CircuitBreaker circuitBreaker) {
             this.circuitBreaker = circuitBreaker;
             return this;
         }